refactor(JOE-977): split postgres-runtime queries catalog and counts
Extract topics/sources/identities catalog reads and table/status count
helpers from queries.ts (~799 → ~615) into queries-catalog.ts and
queries-counts.ts. Keep public re-exports; update module guide and audit.
